Remove Account class that overrides the ActiveRecord one
This class already comes from `spec/active_record_models`. By overriding
it, when all files are loaded, this one ends up creating some kind of
conflict with the ActiveRecord one, and not pushing the values down to
the database as it is required to other tests to work (foreign key and
primary key ones).