Let CustomPrimaryKeyTask class have its own table
No real reason to point to another table just because they share one of
the same attribute: `name`.

It was more confusing to debug because `projects` table has `account_id`
attribute, which is not used and not what CustomPrimaryKeyTask
represents. It's there because ActiveRecord loads the attributes from
the table.